requirement.txt
CherryPy==10.2.2
Flask==0.11.1
Flask-Login==0.4.0
flask-paginate==0.4.6
Jinja2==2.8
PasteScript==2.0.2
mysql-connector==2.1.4
mutagen==1.37
numpy==1.12.1
pandas==0.18.1
librosa==0.5.0
scikit-learn==0.18.1
flask-dropzone==1.3
# external library for read mp3 data
# ffmpeg
# external library
# pyspark
# tensorflow==1.1